US-9713 PS4 boring ssl throwing `SIGPIPE` signal
App "crashes" because it receives a SIGPIPE signal and the default behaviour. To fix this we must use PS4's API to avoid it.

To Fix the issues I added a define `OPENSSL_PS4` to include a header that have wrapper functions for send and receive functions. In sock_read and sock_write I added the appropriate functions for the PS4 `YiNetSend` and `YiNetReceive`. These functions will make sure that the `SIGPIPE` signal doesn't get emitted.