ZD-4162 Fixed TLS Expiration Error
When connecting to a server that has a cert expired in the cert chain
an error of X509_V_ERR_CERT_HAS_EXPIRED. Which in modern implementations
it should not be an issue.

To fix this issue the flag X509_V_FLAG_TRUSTED_FIRST which according to
docs it does:
When X509_V_FLAG_TRUSTED_FIRST is set, construction of the certificate
chain in X509_verify_cert(3) will search the trust store for issuer
certificates before searching the provided untrusted certificates. Local
issuer certificates are often more likely to satisfy local security
requirements and lead to a locally trusted root. This is especially
important when some certificates in the trust store have explicit trust
settings (see "TRUST SETTINGS" in x509(1)). As of OpenSSL 1.1.0 this
option is on by default.

As stated above this is default in openssl(this repos is forked from it)
but in this repo it is not set by default.